test clean up for string const on fail load fails
Resource lists are no longer populated with an error message when a list
could not be found. Instead, we're just leaving one string.empty in the
list and going from there. Updated tests. Reason for this is that we can
more easily write tests against prop fillers, and in reality a resource
list is pretty stable once added to the project.