Add support for GenFu instances. Static usage is still supported.
The only breaking change is the class CalendarDate that doesn't support static usage anymore.

Property fillers are overloaded to allow passing a GenFu instance. They will use the global instance when none is provided to avoid breaking changes, but I have the feeling that requiring an instance would be better for clarity.